test-app/requirements.txt
mysql-connector-python==8.0.28
flask==2.0.3
gunicorn==20.1.0


